requirements.txt
# Quantitative Trading Framework - Dependencies

# Core Data & Computation
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn

# Market Data
yfinance>=0.2.28

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Web UI
streamlit>=1.28.0

# Configuration
PyYAML>=6.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Logging & Utilities
python-dateutil>=2.8.0

# Optional: Live Trading Brokers (install as needed)
# alpaca-trade-api>=3.0.0        # For Alpaca trading
# ib_insync>=0.9.86              # For Interactive Brokers
# python-binance>=1.0.19         # For Binance

# Machine Learning (required for ML features)
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0

# Deep Learning (PyTorch)
torch>=2.0.0
torchvision>=0.15.0

# Optional: Advanced ML/DL (install as needed)
# tensorflow>=2.13.0           # For TensorFlow models (if needed)
# shap>=0.42.0                 # For feature importance analysis

# Optional: Advanced Analytics
# statsmodels>=0.14.0            # For statistical models
# arch>=6.0.0                    # For volatility models
# pyfolio>=0.9.2                 # For portfolio analysis

# Optional: Database Support
# sqlalchemy>=2.0.0
# pymongo>=4.4.0

# Optional: Interactive Dashboards
# plotly>=5.14.0
# dash>=2.11.0
# streamlit>=1.24.0

# Development Tools (optional)
# black>=23.3.0                  # Code formatting
# flake8>=6.0.0                  # Linting
# mypy>=1.3.0                    # Type checking

